boardfarm: prplmesh_base_test: sleep after configure_ssids
After configuring the SSIDs by calling AccessPointCommit, we want to
make use of those SSIDs. In fact, the two tests that currently use
configure_ssids associate a station with the newly configured SSID
immediately afterwards. That should however only be done once the SSID
is really available on the agent.

Ideally, we should wait until the SSIDs are indeed configured, by
polling the agent status. For now, however, just sleep for 5 seconds,
which should be plenty.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>