agent: Insert LAN ports to AgentDB only if they are exist
Since the LAN ports configured on the CMAKE file are exist only on an Axepoint platform,
other platforms that run openWRT would not have these interfaces.
Therefore, do not insert the configured interfaces into the database if they are not exist.
This is a temporary hack so we can run some certification tests on Axepoint.
Need to replace the CMAKE configuration which reflected on "beerocks_agent.conf", to
UCI reading.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>