framework: tlvftypes: add mac_to_array and mac_from_array
In many places, we copy sMacAddr objects to/from uint8_t arrays. The
mac_to_array and mac_from_array reduce the cognitive load of reading
those copy statements.

There was discussion about putting the source or destination first. In
the end, we chose to put the source first, similar to std::copy_n.

Another option was to create a Converter class with to_array and
from_array members. That's a bit more complicated to implement, however,
and the work for mac_to/from_array was already done, so in the end we
went through we that.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>