tree-wide: limit time of generate_connected_clients_events run
Preparative commit.

The generate_connected_clients_events() is used to read the information
on any clients connected to hostapd before the prplmesh was started.
This scenario is most likely to happen on slave_reset() or in prplmesh
processes recovery due to some issue (as well as when we stop and start
prplmesh as part of our tests and our development).
The function generates STA_Connected event for those clients so the
prplmesh will handle them as if they connected after the prplmesh is up
and the prplmesh received STA_Connected event from the hotsapd.

The above operation may take a long time to perform (when many clients
are connected) and this may lead to the monitor thread not handling
events for a long time as well as a burst of events to the monitor
thread and a following burst of notifications from the monitor thread
to the agent, etc.

This commit extends the generate_connected_clients_events() to receive
max-iteration-timeout which will be the timeout of the thread from which
this operation is performed.

This change enables us to move the operation from the Init stage to the
steady-state and defragment the events generations across several thread
awake times.

Signed-off-by: Adam Dov <adamx.dov@intel.com>