common: prplmesh_utils.sh: ignore SIGPIPE and SIGHUP
When the start script is run over SSH, the connection can close before
the agent had a chance to start. When this happens, we get an output
until 'prplmesh_agent_start - start beerocks_agent process...', but
prplMesh doesn't actually start.

This is because prplMesh is started in the background, and the SSH
connection is closed as soon as the main script exits. As a
consequence stdin, stdout and stderr get closed. Any subsequent
read/write from/to them will cause the processes to receive
SIGPIPE (and terminate).

To make sure prplMesh can still start in those cases, explicitly
ignore SIGPIPE. Note that the handler MUST be the empty string for it
to be inherited by the sub-processes (this is because only signal
handlers which are SIG_IGN or SIG_DFL are retained).

To make sure we still continue to start prplMesh in case SIGHUP is
received, also ignore it.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>