agent: Add new BSSID information fields to the AgentDB
Separate the "type" field on the BSSID information on the AgentDB that
represent if the BSS is fBSS or bBSS to two separate flags.
Also, add the backhaul_bss_disallow_profile1/2_agent_association flags,
since on a future commit, we will need to use it not on the
APMANAGER_HOSTAP_VAPS_LIST_UPDATE_NOTIFICATION handler.

Note that this commit breaks compilations to ease code review and it
will be fixed on the next commit.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>