agent: Use new BSSID info data fields
Split the backhaul vap flag into two separate flags on
APMANAGER_HOSTAP_VAPS_LIST_UPDATE_NOTIFICATION message.
On the handler in the son_slave, save it separately on the AgentDB,
and save the "profile1/2_backhaul_sta_association_disallowed" as well.

As part of this change, remove the prepared infrastructure to traffic separation
configuration on the message handler.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>