bcl: Add set_iface_vid_policy() function to network_utils
Preparative commit.
Add function that sets VLAN policy on a specific interface.
This function will be needed on a future commit.

Note that the implementation is based on shell command. This is temporary so we could pass certification on Axepoint quickly.
In the future, we should replace the implementation to be based on Netlink instead.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>