agent: Create VLAN Interface linked to the bridge
Create a VLAN interface linked to the bridge interface for each Secondary VLAN.
For example, if the secondaries VLAN 20 and 30, are configured on the code will add
the interfaces "br-lan.20" and "br-lan.30".

The added code assign a new IP address on a new subnet for each VLAN interface,
bring the interfaces up, and save the subnet and the netmask on a list, as a preparation
for reconfiguring the DHCP server.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>