agent: Add DHCP helper functions
Add two DHCP helper functions:
1. reconf_dhcp() - Reconfigure the DHCP server making it lease
IP addresses for DHCP requests on the new VLANs of the bridge
interfaces. This function should be used on the GW.

2. assign_ip_to_vlan_iface() - Function that sends DHCP request
for assigning an IP address the new VLANs of the bridge
interfaces. This function should be used on the Repeater.

Both functions implemented using shell commands, and they
probably remain that way.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>