agent: Update the son_slave immediately on wireless backhaul connection
Update the son slave when the backhaul connects wirelessly with
newly added BACKHAUL_APPLY_VLAN_POLICY_REQUEST message
so the son_slave will update the traffic separation policy on the Agent.

This is required since the Agent that we connected to should report it's
Primary VLAN and its profile, and that requires VLAN configuration that
without them, the Agent will not be able to send and receive
configuration/topology messages.

The apply of the traffic separation configuration on the new message
handler will be done on the next PR.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>