agent: Apply traffic separation policy on backhaul manager request
Apply traffic separation policy when receiving BACKHAUL_APPLY_VLAN_POLICY_REQUEST
message from the backhaul connects after it connected wirelessly.

This is required since the Agent that we connected to should report it's
Primary VLAN and its profile, and that requires VLAN configuration that
without them, the Agent will not be able to send and receive
configuration/topology messages.

Note that the call to apply_traffic_separation() is without radio argument,
and it means that the policy will only configure the bridge, the wireless
backhaul, and the ethernet ports so the Agent will will configure the Primary
VLAN of the network and be able to get messages from the Controller until
finishing the autoconfiguration process.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>