boardfarm: 4_7_10 simulation
Added phase 6 checks.
This phase wasn't checked up till now as it was
supposed to work "as before". However, certification
tests showed that we are missing Radio Metrics TLV
in AP Metrics Response.
This commit adds triggering report policy and
checks the metrics-response for having the added
radio metrics tlv

Signed-off-by: Ran Regev <ran.regev@devalore.com>