BWL: fix DCS flow
The new DWPAL APIs are synchronous and do not rely on events.
This means that the DWPAL will no longer send out the SCAN_FINISHED
event.
This could cause issues as the DCS flow uses the SCAN_FINISHED to know
when a scan is finished.

To resolve this need to manually send a SCAN_FINISHED event once the
dump sequence is done.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>