channel_scan_task: remove class data member for condition
The channel scan might be triggered externally (iw command or other way)
which may cause the class radio scan to not really exist.

We need to support this by always setting the radio scan to false when
scan-finished is received.

The other task operations are dependant on the task state, thus should
still be performed for cases when the scan was triggered by the task.

Signed-off-by: Adam Dov <adamx.dov@intel.com>