controller: trigger channel scans of next agent after tlv error
The trigger_pending_scan_requests() function is responsible for
building and sending pending scan requests to the agents.

Currently, when the function fails to create or fill a tlv,
it will return and abort the triggering of any pending scans in
other agents.

This commit change the function recovery behavior by continuing to
try and trigger other agents (creating and filling new CMDUs and TLVs),
even when the triggering of the current agent has failed.

PPM-1110

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>