controller: remove log warnings for unexpected scan reports
Since channel scans can be triggered by other processes outside the
dynamic_channel_selection_r2_task, it is possible to receive a channel
scan report for an agent which the task didn't consider as busy.

This commit removes printing of log warnings when a scan report is
received from an unexpected agent, in order to not flood the log with
unneeded reports.

PPM-1110

Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>
Signed-off-by: Daniel Edri <daniel.edri@devalore.com>