tree-wide: stub generate_connected_clients_events return true
It is seen that some of the derived classes of the mon_wlan_hal and
ap_wlan_hal implement the generate_connected_clients_events() API as a
stub and return false. This prevents us for validating the flow for the
return value of the API since the prplmesh CI fails due to the dummy
implementation returning false.

Changed all stub implementations of generate_connected_clients_events()
to return true.

Added a task for implementing the nl80211 API: PPM-1152

Signed-off-by: Adam Dov <adamx.dov@intel.com>