channel_selection_task: release antenna only if zwdfs enabled
The task is always started in switch-antenna off state to release the
antenna on startup in case of recovery from a crash of the agent, but we
must not touch the antenna at all if the feature is not enabled in
configuration - not to interfere in case external daemon manages
the feature.

Skip this case if the feature is disabled and move directly to
not-running.

Signed-off-by: Adam Dov <adamx.dov@intel.com>