controller: topology-task: remove handle dead node for vaps
The topology response handling in the controller was recently changed to
add/remove VAPs according to the received topology information.

As part of that change a call for handle_dead_node() for the removed
VAPs was added which in practice caused a handle_dead_node for the radio
and not the VAP. This is because for VAPs we only remove from the VAP
list and not perform handle_dead_node.

Instead, perform handle_dead_node() for the clients connected to the
removed VAPs and then remove from the DB for that radio.

Signed-off-by: Adam Dov <adamx.dov@intel.com>