tools: openwrt: bump prplwrt to latest main branch
This includes the changes from (and everything that was merged in
prplwrt before that):
https://gitlab.com/prpl-foundation/prplwrt/prplwrt/-/merge_requests/80

Among other things, this implies:

- The prplMesh package is now built by the prpl profile by default, so
we no longer have to add it manually. We can also remove the prplMesh
variant since it's unused (and the variants have anyway been removed
in feed-prpl).

- The 'mmx' and 'sah' are now part of the prplwrt repository, so we
can remove them from our repository.

- We get the fix for libtrace's dependency on liblzma, which allows us
to build with mmx included for 'intel_mips' again.

- The wireless interfaces are now disabled by default for intel_mips.

Closes: https://jira.prplfoundation.org/browse/PPM-1112
Closes: https://jira.prplfoundation.org/browse/PPM-1108

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>