common: prplmesh_utils.sh: agent: read radio names from logfile name
When running the utils script with the 'status' argument, it used to
look for log file with radio names "wlan0" and "wlan2". As a
consequence, the slaves are always reported as not operational on
non-Intel platforms (unless we force them to use those specific
interface names).

In reality, those radio names could be anything (e.g. 'wlan1-1',
'ap1_radio1', ...), so we need a way to find them. We can't simply
read them from UCI, because not all platforms use UCI (for example
rdk-b doesn't use UCI).

To find the filename of the slave logs, look for existing log files in
LOGS_PATH. Rotated log files use names like
'beerocks_agent_wlan0-1.20210216_095241.log', which means we can't
just match anything between 'beerocks_agent_' and '.log' to find the
radio name (we also can't use the dot as a separator, since it's a
valid character for an interface name). Since the most recent log
files are always a symlinks, filter out anything that is not a
symlink. At the same time this will also ensure we exit the for loop
if there is no matching file (in which case it would run once with the
glob turned into a literal asterisk).

Once we have the latest slave logs filenames, we can find the name of
the radio by discarding anything between 'beerocks_agent_' and
'.log' (using BEEROCKS_LOG_FILES_SUFFIX). We can then call
'radio_agent_operational()' as it was done before, with the correct
radio name.

With this approach, there is a potential race: it takes some time for
the log file(s) to be created, and during that time it looks as if
they're operational.

In practice, however, the log files for all slaves will be created
before any of them become operational (the interaction between the
ap_manager and main agent state machines makes sure of that). So it is
sufficient to add a check that at least one operational agent was
found.

While at it, remove the part that displays the log files content. It's
a leftover from a time where we didn't have log files in CI artifacts,
and it's no longer needed.

Closes: https://jira.prplfoundation.org/browse/PPM-1137

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>