tests: boardfarm: remove start_wps_registration from APConfigRenew
APConfigRenew is run both in docker and on the rax40, and it triggers
WPS at the end of the test to verify that the call prints the
corresponding line in the logs.

We only have one boardfarm testbed for now, and it's within Wifi range
of the certification testbed. Since we don't have anything preventing
WPS to be triggered on both testbeds at the same time, we have to
refrain from running any wireless backhaul test in boardfarm for
now (CapiWirelessOnboarding is the only other existing wireless
backhaul test, but it's not run on device at the moment).

Since the test doesn't actually do anything besides triggering WPS,
just remove it for now.

When we decide to add wireless backhaul tests in Boardfarm, it would
probably be a good idea to create separate tests for them
anyway (ideally re-using common code). It would allow one of the two
variants to pass if wired/wireless backhaul doesn't work, and would
also make it possible to run only the wired backhaul tests on testbeds
on which we can't use wireless backhauls.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>