tests: boardfarm: prplmesh_prplwrt: remove prplMesh logs before starting
With copy_logs() Boardfarm now save the logs files from the device.

However, since they are not removed automatically before prplMesh
starts, it means we always save logs from previous runs as well.

Before prplMesh starts, remove the logs on the device.

Because the AlEntity is not yet created before prplMesh starts, we
can't use command() and have to use sendline() directly.

Since copy_logs() also needs the pass to the logs, store it in a class
variable.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>