tests: boardfarm: use a more permissive regex for the prompt
The hostname can be different depending on the device. Some have it
set to "OpenWrt", some to "prplWrt".

Furthermore, it can be convenient for developers to set a custom
hostname to be able to recognize the device more easily.

Accept any hostname that doesn't contain a space.

While at it, also accept paths that are different than '~' or '/', so
long as they don't contain spaces.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>