common: dummy: log when WPS is triggered on the STA
This is a preparatory commit.

Up to now only WPS triggered on the AP was being logged.

Also log when triggering WPS on the STA, to be able to see in the logs
that it was triggered.

Use the same logging line than on other bwl implementations, to make
it possible to check it in Boardfarm.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>