common: bwl: nl80211: better support dual band radios
In 'bfa694bcf887d94bb1e4040dbe3cf3f4fa9476e8' we worked around the
fact that dual band radios are not taken into account in struct
RadioInfo by considering dual bands radios to be 5Ghz radios only.

In practice, this turns out to be a problem for the channel
selection. Since only 5Ghz channels are reported as
"supported_channels" but both 5Ghz and 2.4Ghz channels are reported as
"preferred_channels", unify_channels_list() fails with the following
message:
FATAL 13:19:32:578 <ap_manager> ap_manager_thread.cpp[170] --> ACS reported channel not supproted by the radio: 11

Instead of adding only the channels for the 5Ghz bands, add the
channels of both bands.

For the frequency, check if both 2.4Ghz and 5Ghz are supported, and
use FREQ_24G_5G if it's the case.

For the rest of the m_radio_info members (from max_bandwidth to
vht_mvs_set), we assume the reported parameters will be the same for
both bands and take only one of them into account. A quick check with
'iw' seems to report the same parameters for both bands on a Turris
Omnia. In practice this might not be true, so it has to be verified.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>