agent: Change autoconfiguration message TLV parsing order
Currently, the traffic separation TLVs are being parsed processed before
the RUID comparison to the son_slave RUID.
This is clearly a mistake since the son_slave handles these TLVs even when
the message is directed to the other son_slave.
Move the handling of the traffic separation TLVs to after the RUID comparison.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>