bcl: Modify "is_pid_running()" function
"os_utils::is_pid_running()" function adds to "pid/" to the
path given on its input arguments.
This behavior is not general since it assumes internally that
the path must be on a "pid" folder as we use it only for
beerocks pid files which are located on "/tmp/beerocks/pid/".

Since this function is used by the function "os_utils::kill_pid()"
which we used on one of the previous commits where killing
the PID of the dnsmasq, and the pid file does not located in
a "pid folder, remove the adding of the "pid/" to the given
path, and instead add it externally where needed.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>