tools: add a script to apply UCI settings on a target
This is a preparatory commit.

With the wireless interfaces being disabled by default, we will need
to apply UCI settings for boardfarm, not only for certification tests.

To try to keep the CI yaml files simple (to allow doing the same
operations outside CI, and to make it possible to run linters on
scripts) add a separate bash script to apply UCI settings.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>