ci: use setup_uci.sh to setup the UCI
Replace the shell snippet used to apply the UCI settings by
tools/setup_uci.sh.

'.test-on-target' didn't have it yet. However, since the wireless will
soon be disabled by default on intel_mips as well, we need to start
using it for this job as well.

Finally, for the boardfarm tests we need to apply the same UCI
settings on the netgear-rax40 than on the axepoint. For that, create a
symlink under tests/certification.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>