builder: rdk-b: use 2021Feb27 tag
Up to now, we were just using the "latest" RDK-B release to build for
Turris Omnia. However, this is changing all the time, which leads to
occasional breakage.

Switch to a tag that is hopefully more stable. This is just a snapshot,
so no "better quality" than anything else - but at least, it shouldn't
be modified any more.

There are 3 such tags: RDK Gateway (based on the 2020q4 release, using
OE 2.2), RDK Extender, and Dunfell Gateway (based on the dunfell branch,
i.e. using OE 3.1). Since we were already based on Dunfell, use the
latter one.

See https://wiki.rdkcentral.com/display/RDK/Feb_27_2021_Dunfell_Gateway

Unfortunately, these tags still use the master branch of the RDK and CMF
repositories. Therefore, in local.conf, override the configuration
variables pointing to master with the nightly/20210227-dunfell tag
(well, it's actually a branch, but no commits are made to it).

With this, the RDK-B build succeeds again, so re-enable it in CI.
This reverts commit 0eee5b518d2e92cb63525f63b5834cba209361cd.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>