tools: wait for the device to initialize after a sysupgrade
As described in [1], the devices need some time to initialize and
execute scripts that are only run the first time they boot after an
upgrade.

'afadb39f34a33d3cc82af4ba1ef591a9c4fd840c' already tried to address
the issues that can be caused by changing the UCI configuration when
the device is not initialized yet, but it doesn't work. Remove it, and
just wait for a fixed amount of time after the device has booted
instead.

We already did it for the u-boot-based upgrades, so refactor the
initialization time into a class variable that can be used for
sysupgrade-based upgrades as well.

[1]: https://jira.prplfoundation.org/browse/PPM-1175

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>