docker: builder: rdk-b: save logs as artifacts
When the build fails, there's not much output available for debugging -
just the bitbake error message.

To improve debugging, copy all the log files (the temp directories) to
the artifacts directory after the build. Since this has to be done even
if the build fails, remove the `set -e` and instead add an explicit
error handler for each command that may fail.

While we're at it, rename the artifacts directory inside docker to
/artifacts. This is more in line with how docker normally operates, and
avoids potential problems with ~ expansion.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>