tests: fail when tests are skipped
The command listing skipped tests was using "SKIP" as the grade, but
boardfarm instead set it to null, so fix it.

In addition, add an exit statement with a non-zero code to make sure
the CI job fails when any test is skipped. Since we control what runs
in CI, we're not expecting any skipped test in a normal run.

While at it, fix the indentation and the usage of mixed tabs and
spaces.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>