agent: Solve R1 compatibility issue
If the Controller is Profile-1, it isn't aware of the Profile-2 AP Capability TLV which
the Agent sends on M1 and contains a flag which says whether the Agent supports Combined
bBSS for Profile-1 and Profile-2 Backhaul station connection.
As well, the Profile-1 Controller does not aware of the "Profile-1/2_Backhaul Station
Association Disallowed" bits which were added to the Multi-AP Extention subelement.
If the Agent does not support Combined Profile Backhaul station connection, and a
Profile-1 Controller configures a bBSS, it will send the "Profile-1/2_Backhaul Station
Association Disallowed" set to zero which is interpreted as "Combined Profile bBSS" which
we don't support.
To resolve this, downgrade the bBSS link to Profile-1 Only. This promise that we will
pass R1 Agent certification tests which use Profile-1 Controller.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>