tools: deploy_firmware.sh: retry once if update_fullimage fails
In some cases (see for example [1]) the device is consistently unable
to download the image from the tftp server. When debugging it
manually, it was observed that the only way to make the device able to
download the image again is to reboot it.

To try to avoid this issue in CI, the following changes are done:

- Refactor the update into '_update_fullimage()'.

- Use a try-except to update the first time. If it fails, reboot the
device and try one more time.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1081610024

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>