TLVF: protect against multiple allocation of strings
In the TLVF generated code, when the `set_string()` API is called, there
is space allocation for the new string.
In case this API is called more than once, it may cause allocation of
unnecessary space for the second string.
Since TLVF doesn't really support pointer rollback, this commit just add
protection that the string wasn't already set.

Signed-off-by: Semyon Barenboym <semyon.barenboym@devalore.com>