boardfarm: prplmesh_station: set associated_vap attribute
The common teardown uses the associated_vap attribute of the STA device
object to determine if it needs to be disassociated. This attribute is
however not set on the PrplMeshStation variant.

Set the attribute in the constructor, in wifi_connect and in
wifi_disconnect, like it's done in the dummy variant. Note that the one
in the constructor is redundant since a wifi_disconnect is done
immediately afterwards, but this is more consistent and avoids problems
if the wifi_disconnect is ever removed from the constructor.

Note that the calls to wifi_disconnect now require an argument (since
that's what's expected in the dummy STA as well).

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>