boardfarm: print exception on start
https://jira.prplfoundation.org/browse/PPM-1186

We override the boardfarm `startMarker` to print the topology before
a test begins.

If an exception is raised in this method, it will be silenced and the
test will be skipped with no apparent reason.

Add a broad try-except in the `startMarker` method to print any
exception, so that we have a chance to know why a test was skipped.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>