tlvf: init() must only set dynamic-length index if parsing
For dynamic-length fields (i.e., fields for which the length is
derived from the remaining length of the entire TLV) that belong to a
non-TLV class (e.g. any of the beerocks action classes that are used
in VS messages), the index (= the number of elements in the list) was
always set by the init() function.  However, this should only be done
when parsing, not when creating an empty TLV. In addition, the
end-of-buffer pointer wasn't updated, and the index was set in bytes
instead of in whatever size the field type has.

In fact, the code for this case should be exactly the same as for a TLV
class. The only difference is that the length is derived from the
remaining buffer size instead of the (non-existent) length field.

This issue was not discovered before because:
- this kind of situation only exists for association_frame in
  cACTION_APMANAGER_CLIENT_ASSOCIATED_NOTIFICATION;
- this field is only set with the set_association_frame function;
- that function always writes at the beginning, i.e. doesn't take into
  account the index;
- until commit 866021fa6, that function didn't check that the index is
  indeed 0;
- in commit 866021fa6, all boardfarm tests were skipped so this wasn't
  caught in CI.

Fix this by reusing the complete same code for non-TLV classes as for
TLV classes.

Update AutoGenerated.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>