agent: set number of supported bsses to 4
We currently don't use a real number for number os supported bsses, just
set it to constant 2. One of our new tests, however, wants to configure
more than 2 SSIDs. Since this is anyway just an arbitrary number,
increase it to 3. Note that this corresponds to the number of VAPs
available on MaxLinear devices.

This value was originally brought back from 4 to in in commit e0440ea94
because we would send a teardown M2 for every BSS that doesn't have a
bss_info configuration. However, that has changed since then - no
teardown is sent at all unless no BSSes are configured. So the original
reason to reduce it from 4 has disappeared.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>