tests: openwrt: capture logs outside of the test
Recently the "operational "tests have been failing before
test_status.sh actually started, leaving us unable to understand why
they failed (see for example [1] and [2]).

To make sure we always capture logs, move the capture to a separate
file that is called in the 'after_script' section of the CI job.

Some parts of the log capture have been modified:

- Previously different logs were captured before and after the test,
but the parts that were captured before the test are not affected by
the test. Move everything to after the test.

- 'dmesg' has been replaced by 'logread'. The latter gives us the full
syslog, including the kernel messages.

- In this very limited test, the outputs of the bml_conn_map and
bml_nw_map_query are not useful, so they are simply removed.

- Replace capturing '/etc/*_version' with '/etc/*version', to include
'/etc/prplwrt-version'.

While at it, add copyright headers to both files.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1090896803
[2]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1090849835

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>