docker runner: mount /lib/modules to allow ebtables modprobe
The prplmesh start script will configure ebtables, which requires the
ebtables module to be loaded in the kernel. ebtables will do this
automatically - however, we must make sure that it's possible to load
modules from within the container.

This can be achieved very simply by mounting /lib/modules into the
container.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>