boardfarm: prplmesh_config.json: change iface_to_device to "to-dut"
Up to now, the configuration file had an interface name that is
hardcoded to the one on Cantonica. This has to be renamed to something
that makes more sense. Cantonica anyway already has a udev rule to make
sure the interface name is stable.

In addition, Cantonica now also has an ifupdown rule to create a macvlan
interface, which fixes PPM-1196.

Change the configuration to use that new interface name. The name is now
something more sensible that can easily be used in other test
environments as well: "to-dut".

Fixes PPM-1196.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>