tools: openwrt: bump to get the MMX updates
In [1], MMX was updated, which brought some bug fixes to stable. The
change however did not make it to master, as master uses the MMX
profile from the prplwrt repository directly, and doesn't have an MMX
profile in the prplMesh repository anymore.

Bump OPENWRT_VERSION to the latest prplWrt hash, which will bring us
those MMX changes.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/merge_requests/2220

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>