ci: update easymesh_cert to use Qualcomm sniffer
Note that this change will make some previously-passing-tests fail,
but it will also make it much easier to detect the WPS timing
out (PPM-1102, PPM-1104 and PPM-1105). We agreed on accepting these
newly-failing tests, and fixing them as we find them in PPM-1211.

This depends on the following merge requests:
https://gitlab.com/prpl2/wfa-certification/easymesh_cert/-/merge_requests/78
https://gitlab.com/prpl2/wfa-certification/sniffer_ucc_agent/-/merge_requests/7

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>