agent: Swap location of traffic separation tlvs
handle_multi_ap_policy_config_request() function on sov slave forward the message
to the Monitor thread. To do it, it swaps the cmdu_rx.
The handlers of the traffic separation were located after the swap occurred, which caused to
the data to be read wrongly.
Relocate the traffic separation tlvs to before the swap occurs.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>