bwl: Change hybrid_mode_supported() return value
Change hybrid_mode_supported() return value from false to true.
The reason for that is because it is because the certification test configures bss a hybrid mode,
although the Agent reports it doesn't support it.
This is a workaround commit. A full solution will be to remove completely this function, and set
the traffic separation capabilities differently.
This will be done on a next PR.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>