agent: channel scan task: Add 20mhz function
Currently the channel scan task assumes that the given channel list in
the request contains the primal channels that need to be scanned.
But after additional testing it was discovered that the channel list can
contain central channels for operating classes above 80MHz.

To correctly align the requested channel list to the scanned channel
list a helper function is needed.

Add a function that given the central channel & bandwidth will return
the primal channel list

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>