agent: channel scan task: refactor trigger scan
Since the monitor and subsequently DWPAL accept 20MHz (primary) channels
for their scan, we need to align the channel scan task to that
requirment, but it is possible to also receive non-primary channels,
such as central channels for operating classes above 80MHz.

To do so, when triggering a scan, only 20Mhz channels need to be sent
in the request from ther agent to the monitor.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>