agent: channel scan task: refactor get results
After the scan sequence is finished, the task will retrive all the
stored results for the given request.
But since the task stored the request's channel list in non-primary
channels, and the results are stored in the agent's DB as primary
channels, the get_scan_results_for_request function needs to be
refactored to account for that diffrance.

When retriving the results, search the agent DB using 20Mhz channels.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>