tests: boardfarm: specify tests to run by command line option
Add possibility to specify individual tests to run via
command line option.

* Add --test option to dctest.py.
* Abuse `TEST_SUITE` environment variable passed to docker
  to contain a list of tests if it starts with `TEST_LIST:`.
* Unpack `TEST_SUITE` in `run_bf.sh`.
* Copy entire `boardfarm_plugins` to a temporary folder.
* Create a temporary test suite that contains only the
  specified tests.

This is a quite complicated approach and one may argue
that boardfarm provides a possibility to use `connect`
test suite as an empty one and then add an individual test
with -e switch.
Unfortunately in this case boardfarm adds an implicit
test named `Interact` as well and it fails.
As its name (and source code) implies, it tries to setup
interaction with a human by reading keyboard commands.
It is probably not what is desired in an automated setup.

PPM-1072
-
Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>