bcl: discard misplaced RTM_NEWLINK events
https://jira.prplfoundation.org/browse/PPM-1204

The Axepoint sends message RTM_DELLINK when an interface is removed from
the bridge followed by a RTM_NEWLINK that is misinterpreted as if the
interface had been added to the bridge back again.

To overcome this problem, the implementation of the bridge state
manager has been modified to read the bridge state upon the reception
of a kernel message, to discard misplaced messages and notify true
state change events only.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>