transport: change log level
https://jira.prplfoundation.org/browse/PPM-1204

The transport process is logging the reception of configuration requests
with DEBUG log level. Since the transport process is normally configured
to log only INFO and above, this messages are often missed.

Configuration requests are received a few times only so logging them
will not add too much noise to the logs and these messages are very
interesting for debugging purposes, so log them with INFO instead of
with DEBUG.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>