bpl: add API to get mandatory interfaces list from UCI
Preparative commit.

The mandatory interfaces configuration in UCI is by default empty, but
when defined - the interfaces in the list will not be filtered out from
interfaces list (even if the interface isn't currently available).

This is needed to support platforms where interfaces might not be
available when prplmesh is started (e.g. MaxLinear proprietary repeater
mode where BH connection is managed by an external daemon that removes
the FH and BH interfaces of the radio used for BH connection until BH
connection is established).

The long term solution will add main-agent/unified-agent registration
to interface state-change events and dynamically add (or remove)
son-slaves radio "chains" will be handled in
https://jira.prplfoundation.org/browse/PPM-1220.

This commit adds a BPL API to get the mandatory interfaces from UCI
configurations if exists.

PPM-833.

Signed-off-by: Adam Dov <adamx.dov@intel.com>