transport: remove unnecesary checks
https://jira.prplfoundation.org/browse/PPM-1204

Before calling `activate_interface()` and `deactivate_interface()` we
first check if the socket for the interface has been already created or
not. However, this verification is not performed always.

Inside these functions we are also checking the socket, so any prior
verification is useless.

In order to make code more homogeneus, remove all checks on the socket
before calling these functions and leave the check performed inside.

Also, to know exactly from where these functions were called and why,
add a log message when the interface status changes.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>