ci: re-enable the boardfarm tests on device
Up to now the IP we used to control the device was set on the
interface that we also use for the backhaul. This caused issues when
the interface was added to/removed from the bridge (causing the "no
route to host" issue - see [1]).

What we need to do is to have two separate networks: one to control
the device, and one for the backhaul (if we want to test data
passing). A possible way to do that without requiring additional
hardware is to use an additional vlan-tagged interface on both
ends (the host and the device).

In practice the 192.168.1.1/24 subnet is already used as a control
subnet for the "test-on-device" tests, so we can't easily re-use it
for vlan-tagged traffic (it would break the existing tests).

Instead, since the boardfarm tests on device were anyway disabled, use
a different subnet to control the device. The 192.168.200.1/24 subnet
has been configured on both ends and will be used as the control
network. It should allow us to have a stable connection to the device
even when the data interface (to-dut) is added to/removed from the
bridge.

In practice if we later decide to use two physically different
networks, the same configuration will also work (we just have to
configure the control network to use the same subnet).

Note that if we want to later test data passing between a device and a
docker container, we'll need to use a different IP than the one
defined in wan_ip (since that one uses vlan-tagging).

To make it clear that wan_ip is only a control IP (i.e. not meant for
data) rename it to "control_ip".

Closes https://jira.prplfoundation.org/browse/PPM-1191

[1]: https://jira.prplfoundation.org/browse/PPM-1196

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>