transport: remove event handlers optionally
https://jira.prplfoundation.org/browse/PPM-1204

Method `deactivate_interface()` removes event handlers for the socket
associated to the interface (and closes the socket).

Among others, this method is being called whenever a socket error
occurs. However, since in this case event handlers have already been
removed by the event loop, we are getting an error.

So, when deactivating an interface, removing event handlers must be
optional. And, on socket error, we avoid removing event handlers twice.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>